+++ test/collect_tests.jl
@@ -473,12 +473,12 @@ end
     @test data[6, :id] == 1 && data[6, :weight] ≈ 0.2
 end
 
-@testset "multirun! and different seeds" begin
+@testset "ensemblerun! and different seeds" begin
     model, as!, ms! = Models.schelling(numagents = 1)
     generator(seed) = Models.schelling(;numagents = 1, seed)[1]
     seeds = [1234, 563, 211]
     adata = [:pos]
-    adf, _ = multirun!(generator, as!, ms!, 2; adata, seeds)
+    adf, _ = ensemblerun!(generator, as!, ms!, 2; adata, seeds)
     @test adf[!, :pos] == unique(adf[!, :pos])
     @test sort!(adf[:, :ensemble]) == [1,1,1,2,2,2,3,3,3]
 end